@@ -4,9 +4,12 @@
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-scripts": "1.0.17"
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-redux": "^6.0.0",
+    "react-scripts": "1.0.17",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0"
   },
   "proxy": "http://localhost:8080",
   "scripts": {
