@@ -1,8 +1,21 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import App from './components/App';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux'
+import rootReducer from './reducers';
+import reduxThunk from 'redux-thunk';
 
-ReactDOM.render(<App />, document.getElementById('root'));
-registerServiceWorker();
+const store = createStore(
+  rootReducer,
+  applyMiddleware(reduxThunk)
+);
+
+render(
+  <Provider store={store}>
+    <App />
+  </Provider>, 
+  document.getElementById('root')
+)
+
